requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiohttp-retry==2.8.3
aioice==0.9.0
aiortc==1.9.0
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.7.0
APScheduler==3.11.0
arrow==1.3.0
asyncua==1.1.5
attrs==24.2.0
av==12.3.0
backoff==2.2.1
boto3==1.35.60
botocore==1.35.76
bravado==11.0.3
bravado-core==6.1.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
commonmark==0.9.1
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
Cython==3.0.11
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
efficientnet==1.0.0
efficientnet_pytorch==0.7.1
fastapi==0.110.3
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.55.2
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.10.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
google-crc32c==1.6.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.25.2
huggingface-hub==0.26.5
humanfriendly==10.0
idna==3.7
ifaddr==0.2.0
image-classifiers==1.0.0
imageio==2.36.1
importlib_resources==6.4.5
iniconfig==2.0.0
isoduration==20.11.0
Jinja2==3.1.4
jiter==0.8.0
jmespath==1.0.1
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Keras-Applications==1.0.8
kiwisolver==1.4.7
lazy_loader==0.4
lightning==2.4.0
lightning-utilities==0.11.9
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.3
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
munch==4.0.0
mypy-extensions==1.0.0
neptune==1.13.0
networkx==3.4.2
numpy==1.26.4
nvidia-ml-py==12.560.30
oauthlib==3.2.2
onnxruntime==1.19.2
openai==1.57.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==10.4.0
pluggy==1.5.0
pretrainedmodels==0.7.4
prometheus-fastapi-instrumentator==6.0.0
prometheus_client==0.21.1
propcache==0.2.1
protobuf==5.29.1
psutil==6.1.0
py-cpuinfo==9.0.0
pybase64==1.0.2
pycocotools==2.0.8
pycparser==2.22
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
pydot==2.0.0
pyee==12.1.1
Pygments==2.18.0
PyJWT==2.10.1
pylibsrtp==0.10.0
pyOpenSSL==24.3.0
pyparsing==3.2.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==2.4.0
pytz==2024.2
PyYAML==6.0.2
redis==5.0.8
referencing==0.35.1
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.0.1
roboflow==1.1.49
rpds-py==0.22.3
s3transfer==0.10.4
safetensors==0.4.5
scikit-image==0.24.0
scipy==1.14.1
seaborn==0.13.2
segmentation-models-pytorch==0.3.4
shapely==2.0.6
shellingham==1.5.4
simplejson==3.19.3
six==1.17.0
slack_sdk==3.33.5
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.37.2
structlog==24.4.0
supervision==0.22.0
swagger-spec-validator==3.0.4
sympy==1.13.1
tifffile==2024.9.20
timm==0.9.7
tldextract==5.1.3
tokenizers==0.20.3
torch==2.5.1
torchaudio==2.5.1
torchmetrics==1.6.0
torchvision==0.20.1
tqdm==4.67.1
twilio==9.3.8
typer==0.12.5
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ultralytics-thop==2.0.12
uri-template==1.3.0
urllib3==2.2.3
webcolors==24.11.1
websocket-client==1.8.0
yarl==1.18.3
zxing-cpp==2.2.0
